+++ CHANGELOG.md
@@ -20,6 +20,7 @@ along with Onur. If not, see <https://www.gnu.org/licenses/>.
 - feat: config files discovery 
 - feat: parse one configuration file
 - feat: parse all configuration files found
+- feat: parse config as Config object
 
 ## 0.1.0